# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
  '/api/films':
    post:
      tags:
        - films
      security:
        - cookieAuth: []
      summary: Create a new film
      description: A new film is created by the authenticated user.
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created
        content:
          application/json:
            schema:
              $ref: '../JSON Schemas/film_creation_schema.json'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  $ref: '../JSON Schemas/film_public_response_schema.json'
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/private': 
    get:
      tags:
        - films/private
      security:
        - cookieAuth: []
      summary: Retrieve the private films of the logged-in user 
      description: >-
        The private films of the logged-in user are retrieved. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '../JSON Schemas/film_private_response_schema.json'
                  next:
                    type: string
                    format: uri
                    default: "/api/films/2?pageNo=2"
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {} 
  '/api/films/private/{filmId}':
    get:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Retrieve a private film
      description: The private film with ID filmId is retrieved. This operation can be performed on the film if the user who performs the operation is the film's owner.
      operationId: getSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '../JSON Schemas/film_private_response_schema.json'
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Update a private film
      description: The private film with ID filmId is updated. This operation does not allow changing its visibility. This operation can be performed only by the owner.
      operationId: updateSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '../JSON Schemas/film_update_schema.json'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Delete a private film
      description: The private film with ID filmId is deleted. This operation can only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/public': 
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description: >-
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '../JSON Schemas/film_public_response_schema.json'
                  next:
                    type: "string"
                    format: uri,
                    default: "/api/films/public?pageNo=2"
        '500':
          description: Internal Server Error
          content: {} 
  '/api/films/public/invited': 
    get:
      tags:
        - films/public/invited
      security:
        - cookieAuth: []
      summary: Retrieve the public films that the logged-in user has been invited to review
      description: >-
        The public films that the logged-in user has been invited to review are retrieved. A pagination mechanism is implemented to limit the size of messages.
      operationId: getInvitedFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: ["single", "coop"]
          description: 'The type of review for which the user has been invited'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '../JSON Schemas/film_public_response_schema.json'
                  next:
                    type: string
                    format: uri
        '401':
          description: Unauthorized
          content: {}
        '500':
          description: Internal Server Error
          content: {} 
  '/api/films/public/{filmId}': 
    get:
      tags:
        - films/public/{filmId}
      summary: Retrieve a public film
      description: >-
        The public film with ID filmId is retrieved. This operation does not require authentication.
      operationId: getSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '../JSON Schemas/film_public_response_schema.json'
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
        - 'films/public/{filmId}'
      security:
        - cookieAuth: []
      summary: Update a public film
      description: The public film with ID filmId is updated. This operation does not allow changing its visibility. This operation can be performed only by the owner.
      operationId: updateSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '../JSON Schemas/film_update_schema.json'
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
        - 'films/public/{filmId}'
      security:
        - cookieAuth: []
      summary: Delete a public film
      description: The public film with ID filmId is deleted. This operation can only be performed by the owner.
      operationId: deleteSinglePublicFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/public/{filmId}/reviews':
    post:
      tags:
        - 'films/public/{filmId}/reviews'
      security:
        - cookieAuth: []
      summary: Issue single or cooperative film review to users
      description: The film with ID filmId is assigned to one or more users for review and the corresponding review is created. The users are specified in the reviewers array representation in the request body. This operation can only be performed by the owner.
      operationId: issueFilmReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: the new film reviews, including the users to whom they are issued
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filmId:
                  type: integer
                review_type:
                  type: string
                  default: "coop"
                reviewers:
                  type: array
                  items:
                    type: integer
                    minItems: 1
                    default: 1  
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  type: array
                  items:
                    $ref: '../JSON Schemas/review_response_schema.json'
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (e.g., wrong film id or unknown reviewer id)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    get:
      tags:
        - 'films/public/{filmId}/reviews'
      summary: Retrieve the list of all the reviews that have been issued/completed for a film 
      description: >-
        All the reviews that have been issued/completed for the film with ID filmId are retrieved. A pagination mechanism is implemented to limit the size of messages.
        This operation does not require authentication. 
      operationId: getFilmReviews
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: ID of the requested page (if absent, the first page is returned)'
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: ["single", "coop"]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  reviews:
                    type: array
                    items:
                      $ref: '../JSON Schemas/review_response_schema.json'
                  next:
                    type: string
                    format: uri
                    default: "/api/films/public/2/reviews?pageNo=2"
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/public/{filmId}/reviews/{reviewId}':
    get:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}'
      summary: Retrieve a review that has been issued/completed for a film 
      description: >-
        The review with the given reviewId of the film with ID filmId is retrieved. This operation does not require authentication. 
      operationId: getSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose reviews must be retrieved
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '../JSON Schemas/review_response_schema.json'
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    put:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}'
      security:
        - cookieAuth: []
      summary: Complete a review
      description: The review with ID reviewId of the film with ID filmId is completed. This operation only allows setting the "completed" property to the "true" value, and changing the values of the "reviewDate", "rating", and "review" properties. This operation can be performed only by one of the invited reviewers.
      operationId: updateSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review must be completed
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The properties of the Review object that must be updated
        content:
          application/json:
            schema:
              $ref: '../JSON Schemas/review_update_schema.json'
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    delete:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}'
      security:
        - cookieAuth: []
      summary: Delete a review invitation, either single or cooperative
      description: The review with ID reviewId of the film with ID filmId is deleted. This operation can only be performed by the owner, and only if the review has not yet been completed.
      operationId: deleteSingleReview
      parameters:
        - name: filmId
          in: path
          description: ID of the film whose review invitation must be deleted
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
  '/api/films/public/{filmId}/reviews/{reviewId}/drafts':
    get:
      tags:
        - 'films/public/{filmId}/reviews/{reviewId}/drafts'
      summary: Retrieve all the drafts that have been created for a specific film review
      description: >-
        The drafts of a review with ID reviewId associated to a film with ID filmId are retrieved, This operation can only be formed by users that have been invited to review this film. 
      operationId: getAllDrafts
      parameters:
        - name: filmId
          in: path
          description: ID of the film associated to the review for which the draft will be retrieved
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review whose drafts will be retrieved
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: ID of the requested page (if absent, the first page is returned)'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  reviews:
                    type: array
                    items:
                      $ref: '../JSON Schemas/draft_response_schema.json'
                  next:
                    type: string
                    format: uri
                    default: "/api/films/public/2/reviews/1/drafts?pageNo=2"
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '500':
          description: Internal Server Error
          content: {}
    post:
      tags: 
        - 'films/public/{filmId}/reviews/{reviewId}/drafts'
      security:
        - cookieAuth: []
      summary: Create a Draft for a specific review with ID reviewId
      description: A new draft is created for the review with ID reviewId. The author of the draft is specified in the body. This operation can only be performed for cooperative reviews by an invited reviewer.
      operationId: createReviewDraft
      parameters:
        - name: filmId
          in: path
          description: ID of the film
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          description: ID of the review
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: the new review draft
        content:
          application/json:
            schema:
              type: object
              $ref: '../JSON Schemas/draft_create_schema.json'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  type: array
                  items:
                    $ref: '../JSON Schemas/draft_response_schema.json'
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '409':
          description: Conflict (e.g., wrong film id or unknown reviewer id)
          content: {}
        '500':
          description: Internal Server Error
          content: {}
...
  '/api/users':
    get:
      tags:
        - users
      security:
        - cookieAuth: []
      summary: Get information about the users
      description: >-
        The available information (passwords excluded) about all the
        users is retrieved. This operation is available only to authenticated users.
      operationId: getUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '../JSON Schemas/user_public_schema.json'
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not Found
          content: {}
  '/api/users/{userId}':
    get:
      tags:
        - 'users/{userId}'
      security:
        - cookieAuth: []
      summary: Get information about a user
      description: >-
        The available information (password excluded) about the user
        specified by userId is retrieved.
      operationId: getSingleUser
      parameters:
        - name: userId
          in: path
          description: ID of the user to get
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '../JSON Schemas/user_public_schema.json'
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not Found
          content: {}
  '/api/users/authenticator':
    post:
      tags:
        - users/authenticator
      summary: Logs a user in or out
      description: >-
        The user who wants to log in or out sends the user data to the
        authenticator which performs the operation. For logout, the operation is possible only for authenticated users.
      operationId: authenticateUser
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: The operation type ("login" or "logout")
      requestBody:
        description: >-
          The data of the user who wants to perform log in. For login
          the structure must contain email and password.
        content:
          application/json:
            schema:
              $ref: '../JSON Schemas/user_login_schema.json'
        required: true
      responses:
        '204':
          description: No Content
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
        '400':
          description: Bad Request
          content: {}
        '401':
          description: Unauthorized
          content: {}
components:
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid